Keep mutation coordination outside the domain layer
