+++ src/entities/card/index.ts
@@ -1,6 +1,11 @@
-export { CardBulkMutationError, cardCommands } from "./api/commands";
-export { generateCardId } from "./api/firestore";
-export { removeForDeck as removeCardDocumentsForDeck } from "./api/firestore";
+export {
+  create as createCardDocument,
+  generateCardId,
+  logicalRemove as removeCardDocument,
+  removeForDeck as removeCardDocumentsForDeck,
+  update as updateCardDocument,
+  upsert as upsertCardDocument,
+} from "./api/firestore";
 export { createCard } from "./model/card";
 export type { Card, CardEdit, CardId, CardRaw } from "./model/card";
 export { useCards } from "./hooks/useCards";
