+++ src/features/deck/create/api/createDeck.ts
@@ -1,6 +1,7 @@
 import type { Deck } from "@/entities/deck";
 
-import { createDeckDocument, deckMutationLock } from "@/entities/deck";
+import { createDeckDocument } from "@/entities/deck";
+import { resourceKey } from "@/shared/lib/resourceAccess";
 import { runSerially } from "@/shared/lib/runSerially";
 import { waitForRemoteWrite } from "@/shared/lib/remoteWrite";
 
@@ -11,7 +12,7 @@ const requireOwner = (uid: string, deck: Deck) => {
 
 export const createDeck = async (uid: string, deck: Deck): Promise<void> => {
   requireOwner(uid, deck);
-  await runSerially(deckMutationLock(uid, deck.id), () =>
+  await runSerially(resourceKey("deck", uid, deck.id), () =>
     waitForRemoteWrite(createDeckDocument(deck), "Deck creation")
   );
 };
